+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.2.5
+
+- Add support for cross-batch individual source asset filtering.
+- Add support for cross-batch batch source asset filtering by tag.
+- Add button to set the batch as the reference batch in the batch info panel.
+- Add link to view only two batches in the batch info panel.
+
 ## v0.2.4
 
 - Add the megapixels field.